requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
transformers
datasets
pillow
joblib
fastapi
uvicorn
requests
psycopg2-binary
redis
prometheus-client
python-multipart
pydantic
pydantic-settings
aioredis
tenacity
streamlit